add gate for questions

+++ app/Http/Controllers/QuestionsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Question;
 use App\Http\Requests\AskQuestionRequest;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 
 class QuestionsController extends Controller
@@ -72,7 +73,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        if(Gate::allows('update-question', $question)){
+            return view('questions.edit', compact('question'));
+        }
+
+        abort(403, 'Access denied');
     }
 
     /**
@@ -84,6 +89,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(Gate::denies('update-question', $question))
+        {
+            abort(403, 'Access denied');
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', 'Question has been updated');
@@ -97,6 +107,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question) // $id
     {
+        if(Gate::denies('delete-question', $question))
+        {
+            abort(403, 'Access denied');
+        }
+
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', 'Questions has been deleted!');
